Agent: Move ControlClient.find_server in network/relay/utls.py
